+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [#6] - Support SSG-EG edit (thanks [@freq-mod])
 - Bank support
 - FM/PSG switch key
-- [#3], [#5], [#7] - Support 10 new format import/export (thanks [@freq-mod])
+- [#3], [#5], [#7] - Support 11 new format import/export (thanks [@freq-mod])
   - .bti/.btb (BambooTracker instrument/bank)
   - .ff (PMD FF)
   - .opni/.wopn (OPN2BankEditor imstrument/bank) (import only)
@@ -14,6 +14,7 @@
   - .tfi (TFM Music Maker instrument)
   - .vgi (VGM Music Maker instrument)
   - .y12 (Gens KMod dump) (import only)
+  - .ins (MVSTracker instrument)
 - [#6] - Add Nuked OPN-Mod emulation (thanks [@freq-mod])
 - [#8] - Tone search box (thanks [@freq-mod])